#750 Long values binding fix
Former-commit-id: 61ce94f2c6f95833eab0d9a7d3c62b06b5c0cf7b

+++ plugins/org.jkiss.dbeaver.model/src/org/jkiss/dbeaver/model/impl/jdbc/data/handlers/JDBCNumberValueHandler.java
@@ -193,7 +193,7 @@ public class JDBCNumberValueHandler extends JDBCAbstractValueHandler {
                     break;
                 case java.sql.Types.NUMERIC:
                     if (number instanceof Long) {
-                        statement.setLong(paramIndex, number.intValue());
+                        statement.setLong(paramIndex, number.longValue());
                     } else if (number instanceof Integer) {
                         statement.setInt(paramIndex, number.intValue());
                     } else if (number instanceof Short) {
